requirements.txt
asgiref==3.8.1
channels==4.1.0
daphne==4.1.2
Django==6.0.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0